requirements.txt
parsel==1.8.1
playwright==1.35.0
pre-commit==3.3.3
pytest==7.4.0
python-dotenv==1.0.0
twilio==8.4.0
